Remove metric suffix
I hadn't really considered the implications on counters that having a
suffix would have. At the moment we mandate that counter names are
suffixed with `_total`, so having some suffix that is applied to
every metric name completely breaks that.

I don't think it's worth doing anything clever where `_total` is moved
to the end of the counter name, it just feels like we're asking for
trouble.